+++ stroom-integrationtest/src/integrationTest/java/stroom/pipeline/task/TranslationTest.java
@@ -262,6 +262,8 @@ public abstract class TranslationTest extends AbstractCoreIntegrationTest {
                 }
 
                 // Make sure we have at least one processed stream else it indicates an error in processing somewhere
+                // If we get an error stream you can just run the pipeline in stroom, to try and diagnose the fault
+                // if the above error stream dump doesn't help
                 Assert.assertTrue(processedStreams.size() > 0);
 
                 // Copy the contents of the latest written stream to the output.