+++ stroom-security/stroom-security-server/src/main/java/stroom/security/SecurityFilter.java
@@ -208,7 +208,7 @@ public class SecurityFilter implements Filter {
             LOGGER.debug("We have the following state: {{}}", stateId);
 
             // Check the state is one we requested.
-            final AuthenticationState state = AuthenticationStateSessionUtil.pop(request.getSession(false));
+            final AuthenticationState state = AuthenticationStateSessionUtil.pop(request);
             if (state == null) {
                 LOGGER.warn("Unexpected state: " + stateId);
 
@@ -253,7 +253,7 @@ public class SecurityFilter implements Filter {
         final String url = request.getRequestURL().toString();
 
         // Create a state for this authentication request.
-        final AuthenticationState state = AuthenticationStateSessionUtil.create(request.getSession(true), url);
+        final AuthenticationState state = AuthenticationStateSessionUtil.create(request, url);
 
         // If we're using the request URL we want to trim off any trailing params
         final URI parsedRequestUrl = UriBuilder.fromUri(url).build();